+++ resources/stylesheets/style.css
@@ -9,8 +9,6 @@
   --cities-body-bg-color: #57a;
 }
 
-/* #27f; */
-
 * {
   box-sizing: border-box;
   -ms-overflow-style: none;
@@ -46,26 +44,18 @@ main {
   background-color: var(--main-background-color);
   width: 100%;
   height: calc(100% - var(--header-height));
-  /* border-top: 4px solid var(--header-bg-color);
-  border-bottom: 4px solid var(--header-bg-color);
-  border-right: 4px solid var(--header-bg-color); */
   border: 4px solid var(--header-bg-color);
   display: flex;
 }
 
 #citiesTableView {
-  /* position: absolute;
-  left: 0;
-  top: 0;
-  width: 30%; */
   position: relative;
   border-right: 1px solid #cee;
   flex: 1;
-  /* min-width: 200px; */
   height: 100%;
   border-right: 4px solid var(--header-bg-color);
-  /* border-right: 4px solid var(--header-bg-color); */
   color: white;
+  font-family: Arial, Helvetica, sans-serif;
 }
 
 #citiesTableHeader,
@@ -75,12 +65,8 @@ main {
   top: 0;
   width: 100%;
   height: var(--cities-header-height);
-  /* background-color: var(--cities-header-bg-color); */
   border-bottom: 4px solid var(--header-bg-color);
   font-size: 28px;
-  /* display: flex;
-  justify-content: center;
-  align-items: center; */
 }
 
 #weatherViewHeader {
@@ -93,7 +79,6 @@ main {
 
 #weatherCityName {
   font-size: 36px;
-  /* background-color: #555; */
   margin: 0;
   flex: 4;
   display: flex;
@@ -104,6 +89,7 @@ main {
   flex: 4;
   font-size: 16px;
   font-variant: small-caps;
+  font-variant-caps: all-small-caps;
 }
 
 #searchCity,
@@ -113,7 +99,6 @@ main {
   display: flex;
   justify-content: center;
   align-items: center;
-  /* background-color: #888; */
 }
 
 
@@ -161,15 +146,14 @@ main {
 .citiesTabsBar {
   width: 100%;
   height: 76px;
-  /* background-color: #666; */
   display: flex;
   justify-content: center;
   align-items: center;
   font-variant: small-caps;
+  font-variant-caps: all-small-caps;
 }
 
 .citiesTab {
-  /* width: 30%; */
   flex: 1;
   margin: 30px 5px 0 5px;
   padding: 10px;
@@ -209,7 +193,6 @@ main {
   top: var(--cities-header-height);
   width: 100%;
   height: calc(100% - var(--cities-header-height));
-  /* background-color: var(--cities-body-bg-color); */
   overflow: auto;
   padding: 0 10px 10px 10px;
 }
@@ -240,22 +223,11 @@ main {
 }
 
 #weatherView {
-  /* overflow: auto;
-  position: absolute;
-  left: 30%;
-  width: 70%;
-  height: 100%; */
   position: relative;
   display: block;
   flex: 2;
-  /* background-color: #57a; */
-  /* overflow: auto; */
   padding: 0 10px;
   color: white;
-  /* border: ; */
-  /* border: 4px solid var(--header-bg-color); */
-  /* border-left: none; */
-
 }
 
 #weatherViewBodyTopPanel {
@@ -265,15 +237,12 @@ main {
   display: flex;
   justify-content: center;
   align-items: center;
-  /* background-color: rgba(255, 255, 255, 0.3); */
 }
 
 #weatherViewTemperaturePanel {
-  /* width: 30vw; */
   padding: 0 40px;
   height: 100%;
   font-size: 14vw;
-  /* background-color: rgba(50, 50, 50, 0.3); */
   display: flex;
   justify-content: center;
   align-items: center;
@@ -282,7 +251,6 @@ main {
 #weatherViewTemperaturePanel > span {
   font-size: 2vw;
   margin-bottom: -7vw;
-  /* line-height: 12vw; */
 }
 
 #temperatureToggle {
@@ -353,17 +321,20 @@ main {
 #minContainer {
   justify-content: flex-start;
   font-variant: small-caps;
+  font-variant-caps: all-small-caps;
 }
 
 #maxContainer {
   justify-content: flex-end;
   font-variant: small-caps;
+  font-variant-caps: all-small-caps;
 }
 
 #weatherMain,
 #weatherDescription  {
   justify-content: center;
   font-variant: small-caps;
+  font-variant-caps: all-small-caps;
 }
 
 #weatherBackground {
@@ -372,7 +343,6 @@ main {
   left: 0;
   width: 100%;
   height: 100%;
-  /* background-color: rgba(255, 255, 255, 0.3); */
   background-position: center center;
   background-repeat: no-repeat;
   background-size: 100% 100%;
@@ -385,13 +355,38 @@ main {
   height: 30px;
   bottom: 0;
   left: calc(50% - 110px);
-  background-color: #777;
+  background-color: #aaa;
   text-align: center;
   display: flex;
   justify-content: center;
   align-items: center;
   color: rgb(10, 50, 90);
   font-family: Arial, Helvetica, sans-serif;
+  border-radius: 4px 4px 0 0;
+}
+
+.sunShine {
+  background-image: url('../images/sunshine.jpg');
+}
+
+.clouds {
+  background-image: url('../images/clouds.jpg');
+}
+
+.rain {
+  background-image: url('../images/rain.jpg');
+}
+
+.haze {
+  background-image: url('../images/haze.png');
+}
+
+.snow {
+  background-image: url('../images/snow.jpg');
+}
+
+.fog {
+  background-image: url('../images/fog.jpg');
 }
 
 @media screen and (max-width: 768px) {
